Change the name of sdlog file in case of differential gps
The fix_type variable in ubx.cpp may be assigned to 4(DGPS) or 5(Float RTK) or 6(Fixed RTK). So if we use differential gps, the value of the fix_type variable may not be 3. But gps still has 3d data and gps time. So The name of the file should named with gps time.My ublox differential gps fix_type is 5 (Float RTK).